+++ assets/javascripts/kitten/components/tables/list-table/components/body.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import styled from 'styled-components'
+
+export const ListTableBody = styled.ul`
+  list-style: none;
+  margin: 0;
+  padding: 0;
+`