+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: default `ListTable` column style to Flex.
+
 ## [2.19.0] - 2019-08-19
 
 Features: