+++ CHANGELOG.md
@@ -4,7 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Fix: default `ListTable` column style to Flex.
+## [2.19.1] - 2019-08-20
+
+Fix:
+- Default `ListTable` column style to Flex.
 
 ## [2.19.0] - 2019-08-19